# express-swagger-ui/swagger.yaml
openapi: 3.0.0

info:
  version: '1.0.0'
  title: App axample swagger docs.
  description: We wanna show swagger examples.

paths:
  /user:
    get:
      summary: Returns list of users
      description: Returns list of users with some data.
      operationId: getUsers
      parameters:
        - name: limit
          in: query
          description: Max count of users
          required: false
          schema:
            type: number
        - name: offset
          in: query
          description: Max count of users
          required: false
          schema:
            type: number
      responses:
        '200':
          description: Successful request with list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          description: Invalid limit or offset supplied.
        '404':
          description: No users found.
        '500':
          description: Server internal error.
    post:
      security:
        - bearerAuth: []
      summary: Create new users
      description: Create new users with some data.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              required: true
              properties:
                name:
                  type: string
                  required: true
                surname:
                  type: string
      responses:
        '200':
          description: Successful request with new user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid body params supplied.
        '500':
          description: Server internal error.

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      required:
        - user_id
        - name
        - status
      properties:
        user_id:
          type: string
          format: uuid
        name:
          type: string
        surname:
          type: string
        status:
          type: string
          description: User Status
          enum:
            - active
            - disabled
            - banned
            - deleted

    Users:
      type: array
      items:
        $ref: '#/components/schemas/User'